Wydzielony Product z OfferItem
