+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/OfferItem.java
@@ -19,13 +19,8 @@ import java.math.BigDecimal;
 import java.util.Date;
 
 public class OfferItem {
-
-    // product
-    private String productId;
-    private BigDecimal productPrice;
-    private String productName;
-    private Date productSnapshotDate;
-    private String productType;
+    
+    private Product product;
     private int quantity;
     private BigDecimal totalCost;
     private String currency;
@@ -34,20 +29,11 @@ public class OfferItem {
     private String discountCause;
     private BigDecimal discount;
 
-    public OfferItem(String productId, BigDecimal productPrice, String productName,
-            Date productSnapshotDate, String productType, int quantity) {
-        this(productId, productPrice, productName, productSnapshotDate, productType, quantity, null, null);
+    public OfferItem(int quantity) {
+        this(quantity, null, null);
     }
 
-    public OfferItem(String productId, BigDecimal productPrice, String productName,
-            Date productSnapshotDate, String productType, int quantity,
-            BigDecimal discount, String discountCause) {
-        this.productId = productId;
-        this.productPrice = productPrice;
-        this.productName = productName;
-        this.productSnapshotDate = productSnapshotDate;
-        this.productType = productType;
-
+    public OfferItem(int quantity, BigDecimal discount, String discountCause) {
         this.quantity = quantity;
         this.discount = discount;
         this.discountCause = discountCause;
@@ -61,24 +47,8 @@ public class OfferItem {
                 .multiply(new BigDecimal(quantity)).subtract(discountValue);
     }
 
-    public String getProductId() {
-        return productId;
-    }
-
-    public BigDecimal getProductPrice() {
-        return productPrice;
-    }
-
-    public String getProductName() {
-        return productName;
-    }
-
-    public Date getProductSnapshotDate() {
-        return productSnapshotDate;
-    }
-
-    public String getProductType() {
-        return productType;
+    public String getProduct() {
+        return product;
     }
 
     public BigDecimal getTotalCost() {
