+++ src/main/java/pl/com/bottega/ecommerce/sales/domain/offer/Product.java
@@ -1,0 +1,57 @@
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+package pl.com.bottega.ecommerce.sales.domain.offer;
+
+import java.math.BigDecimal;
+import java.util.Date;
+
+/**
+ *
+ * @author 195059
+ */
+public class Product {
+
+    private String productId;
+    private BigDecimal productPrice;
+    private String productName;
+    private Date productSnapshotDate;
+    private String productType;
+
+    public Product(String productId, BigDecimal productPrice, String productName,
+            Date productSnapshotDate, String productType) {
+        this(productId, productPrice, productName, productSnapshotDate, productType);
+    }
+
+    public Product(String productId, BigDecimal productPrice, String productName,
+            Date productSnapshotDate, String productType) {
+        this.productId = productId;
+        this.productPrice = productPrice;
+        this.productName = productName;
+        this.productSnapshotDate = productSnapshotDate;
+        this.productType = productType;
+    }
+
+    public String getProductId() {
+        return productId;
+    }
+
+    public BigDecimal getProductPrice() {
+        return productPrice;
+    }
+
+    public String getProductName() {
+        return productName;
+    }
+
+    public Date getProductSnapshotDate() {
+        return productSnapshotDate;
+    }
+
+    public String getProductType() {
+        return productType;
+    }
+
+}
